Preview cardinality changes in `pull` output
This PR adds additional output to the `pull` action. Specifically, it will now
generate github action outputs for the total series count and series count
change in each segment.

It will also generate a handy summary in the form of a table. This makes it
easier to review previous recommendation pulls to audit what has changed in
terms of series count.